Untracking protocol buffer stuff, and some minor warning resolutions.
Also removes VS dependencies on protocol buffer stuff from the VS projects, and adjusted the Unix makefile.

+++ unisim/unisim/bson.hpp
@@ -259,7 +259,7 @@ public:
         }
     }
 
-    bool push_string(char* name, char* v, int32_t len = -1, int8_t type = BSONReader::ElementTypes::String)
+    bool push_string(char* name, char* v, size_t len = -1, int8_t type = BSONReader::ElementTypes::String)
     {
         if (child != NULL)
         {
@@ -284,7 +284,7 @@ public:
                 pos += 1;
                 memcpy(out + pos, name, name_len);
                 pos += name_len;
-                *(int32_t*)(out + pos) = len + 1; // string + 0x00
+                *(int32_t*)(out + pos) = (int32_t)(len + 1); // string + 0x00
                 pos += 4;
                 memcpy(out + pos, v, len);
                 pos += len;
@@ -383,7 +383,7 @@ public:
         {
             enlarge(1);
             out[pos] = 0;
-            *(int32_t*)out = pos + 1; // Set the first 4 bytes to the length
+            *(int32_t*)out = (int32_t)(pos + 1); // Set the first 4 bytes to the length
             return out;
         }
     }
@@ -411,11 +411,11 @@ private:
     uint8_t* out;
     int32_t is_array = -1;
     char array_name[10];
-    int32_t pos;
+    uint64_t pos;
     BSONWriter* child;
     BSONWriter* parent;
 
-    void enlarge(int32_t nbytes)
+    void enlarge(size_t nbytes)
     {
         uint8_t* out_new = (uint8_t*)realloc(out, pos + nbytes);
         if (out_new == NULL)
@@ -430,6 +430,7 @@ private:
 
     char* print_array_name()
     {
+        // Windows notes that sprintf_s is available.
         sprintf(array_name, "%d", is_array);
         is_array += 1;
         return array_name;
@@ -437,53 +438,3 @@ private:
 };
 
 #endif
-
-////char msg[] = { '\x31', '\x00', '\x00', '\x00', '\x04', 'B', 'S', 'O', 'N', '\x00', '\x26', '\x00', '\x00', '\x00', '\x02', '\x30', '\x00', '\x08', '\x00', '\x00', '\x00', 'a', 'w', 'e', 's', 'o', 'm', 'e', '\x00', '\x01', '\x31', '\x00', '\x33', '\x33', '\x33', '\x33', '\x33', '\x33', '\x14', '\x40', '\x10', '\x32', '\x00', '\xc2', '\x07', '\x00', '\x00', '\x00', '\x00' };
-//char msg[] = { '\x8b', '\x01', '\x00', '\x00', '\x03', '\x67', '\x6c', '\x6f', '\x73', '\x73', '\x61', '\x72', '\x79', '\x00', '\x7c', '\x01', '\x00', '\x00', '\x03', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x44', '\x69', '\x76', '\x00', '\x51', '\x01', '\x00', '\x00', '\x03', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x4c', '\x69', '\x73', '\x74', '\x00', '\x34', '\x01', '\x00', '\x00', '\x03', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x45', '\x6e', '\x74', '\x72', '\x79', '\x00', '\x23', '\x01', '\x00', '\x00', '\x03', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x44', '\x65', '\x66', '\x00', '\x81', '\x00', '\x00', '\x00', '\x04', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x53', '\x65', '\x65', '\x41', '\x6c', '\x73', '\x6f', '\x00', '\x1b', '\x00', '\x00', '\x00', '\x02', '\x30', '\x00', '\x04', '\x00', '\x00', '\x00', '\x47', '\x4d', '\x4c', '\x00', '\x02', '\x31', '\x00', '\x04', '\x00', '\x00', '\x00', '\x58', '\x4d', '\x4c', '\x00', '\x00', '\x02', '\x70', '\x61', '\x72', '\x61', '\x00', '\x49', '\x00', '\x00', '\x00', '\x41', '\x20', '\x6d', '\x65', '\x74', '\x61', '\x2d', '\x6d', '\x61', '\x72', '\x6b', '\x75', '\x70', '\x20', '\x6c', '\x61', '\x6e', '\x67', '\x75', '\x61', '\x67', '\x65', '\x2c', '\x20', '\x75', '\x73', '\x65', '\x64', '\x20', '\x74', '\x6f', '\x20', '\x63', '\x72', '\x65', '\x61', '\x74', '\x65', '\x20', '\x6d', '\x61', '\x72', '\x6b', '\x75', '\x70', '\x20', '\x6c', '\x61', '\x6e', '\x67', '\x75', '\x61', '\x67', '\x65', '\x73', '\x20', '\x73', '\x75', '\x63', '\x68', '\x20', '\x61', '\x73', '\x20', '\x44', '\x6f', '\x63', '\x42', '\x6f', '\x6f', '\x6b', '\x2e', '\x00', '\x00', '\x02', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x53', '\x65', '\x65', '\x00', '\x07', '\x00', '\x00', '\x00', '\x6d', '\x61', '\x72', '\x6b', '\x75', '\x70', '\x00', '\x02', '\x41', '\x63', '\x72', '\x6f', '\x6e', '\x79', '\x6d', '\x00', '\x05', '\x00', '\x00', '\x00', '\x53', '\x47', '\x4d', '\x4c', '\x00', '\x02', '\x47', '\x6c', '\x6f', '\x73', '\x73', '\x54', '\x65', '\x72', '\x6d', '\x00', '\x25', '\x00', '\x00', '\x00', '\x53', '\x74', '\x61', '\x6e', '\x64', '\x61', '\x72', '\x64', '\x20', '\x47', '\x65', '\x6e', '\x65', '\x72', '\x61', '\x6c', '\x69', '\x7a', '\x65', '\x64', '\x20', '\x4d', '\x61', '\x72', '\x6b', '\x75', '\x70', '\x20', '\x4c', '\x61', '\x6e', '\x67', '\x75', '\x61', '\x67', '\x65', '\x00', '\x02', '\x41', '\x62', '\x62', '\x72', '\x65', '\x76', '\x00', '\x0e', '\x00', '\x00', '\x00', '\x49', '\x53', '\x4f', '\x20', '\x38', '\x38', '\x37', '\x39', '\x3a', '\x31', '\x39', '\x38', '\x36', '\x00', '\x02', '\x53', '\x6f', '\x72', '\x74', '\x41', '\x73', '\x00', '\x05', '\x00', '\x00', '\x00', '\x53', '\x47', '\x4d', '\x4c', '\x00', '\x02', '\x49', '\x44', '\x00', '\x05', '\x00', '\x00', '\x00', '\x53', '\x47', '\x4d', '\x4c', '\x00', '\x00', '\x00', '\x02', '\x74', '\x69', '\x74', '\x6c', '\x65', '\x00', '\x02', '\x00', '\x00', '\x00', '\x53', '\x00', '\x00', '\x02', '\x74', '\x69', '\x74', '\x6c', '\x65', '\x00', '\x11', '\x00', '\x00', '\x00', '\x65', '\x78', '\x61', '\x6d', '\x70', '\x6c', '\x65', '\x20', '\x67', '\x6c', '\x6f', '\x73', '\x73', '\x61', '\x72', '\x79', '\x00', '\x00', '\x00' };
-//BSONReader b(msg);
-//struct BSONReader::Element el = { 0 };
-//while (el.type != -1)
-//{
-//    el = b.get_next_element();
-//}
-//
-//BSONWriter bw;
-////bw.push_array("BSON");
-////bw.push_string("", "awesome", 7);
-////bw.push_double("", 5.05);
-////bw.push_int32("", 1986);
-////bw.push_end();
-//bw.push_subdoc("glossary");
-//bw.push_subdoc("GlossDiv");
-//bw.push_subdoc("GlossList");
-//bw.push_subdoc("GlossEntry");
-//bw.push_subdoc("GlossDef");
-//bw.push_array("GlossSeeAlso");
-//bw.push_string("", "GML");
-//bw.push_string("", "XML");
-//bw.push_end();
-//bw.push_string("para", "A meta-markup language, used to create markup languages such as DocBook.");
-//bw.push_end();
-//bw.push_string("GlossSee", "markup");
-//bw.push_string("Acronym", "SGML");
-//bw.push_string("GlossTerm", "Standard Generalized Markup Language");
-//bw.push_string("Abbrev", "ISO 8879:1986");
-//bw.push_string("SortAs", "SGML");
-//bw.push_string("ID", "SGML");
-//bw.push_end();
-//bw.push_end();
-//bw.push_string("title", "S");
-//bw.push_end();
-//bw.push_string("title", "example glossary");
-//bw.push_end();
-//uint8_t* out = bw.push_end();
-//for (int i = 0; i < *(int32_t*)out; i++)
-//{
-//    printf("%02X %02X\n", out[i], (uint8_t)msg[i]);
-//    //if (msg[i] != out[i])
-//    //{
-//    //    printf("NOPE\n");
-//    //    return 1;
-//    //}
-//}
-//return 0;
